requirements.txt
pygame-ce>=2.5.0
numpy>=1.24.0
Pillow>=10.0.0
librosa>=0.10.0
soundfile>=0.12.0
customtkinter>=5.2.0
moviepy>=1.0.3
PyOpenGL>=3.1.6
proglog>=0.1.10
